Stub out macros and port tests
