+++ cask/src-0/cask/router/RoutesEndpointMetadata.scala
@@ -1,0 +1,9 @@
+package cask.router
+
+import cask.router.EntryPoint
+
+case class RoutesEndpointsMetadata[T](value: EndpointMetadata[T]*)
+object RoutesEndpointsMetadata{
+
+  implicit def initialize[T]: RoutesEndpointsMetadata[T] = ???
+}
